+++ packages/flame_forge2d/example/lib/tappable_sample.dart
@@ -1,26 +1,26 @@
-import 'package:flame_forge2d/flame_forge2d.dart';
-import 'package:forge2d/forge2d.dart';
-import 'package:flame_forge2d/forge2d_game.dart';
-import 'package:flame/palette.dart';
 import 'package:flame/components.dart';
+import 'package:flame/palette.dart';
+import 'package:flame_forge2d/flame_forge2d.dart';
+import 'package:flame_forge2d/forge2d_game.dart';
+import 'package:forge2d/forge2d.dart';
 
 import 'balls.dart';
 import 'boundaries.dart';
 
-class TapableSample extends Forge2DGame with HasTapableComponents {
-  TapableSample() : super(zoom: 20, gravity: Vector2(0, -10.0));
+class TappableSample extends Forge2DGame with HasTappableComponents {
+  TappableSample() : super(zoom: 20, gravity: Vector2(0, -10.0));
 
   @override
   Future<void> onLoad() async {
     final boundaries = createBoundaries(this);
     boundaries.forEach(add);
     final center = screenToWorld(viewport.effectiveSize / 2);
-    add(TapableBall(center));
+    add(TappableBall(center));
   }
 }
 
-class TapableBall extends Ball with Tapable {
-  TapableBall(Vector2 position) : super(position) {
+class TappableBall extends Ball with Tappable {
+  TappableBall(Vector2 position) : super(position) {
     originalPaint = BasicPalette.white.paint();
     paint = originalPaint;
   }